Move //third_party/tensorflow/compiler/xla/service:hlo_live_range{_test} to //third_party/tensorflow/compiler/xla/hlo/utils:hlo_live_range{_test} and update all users.
PiperOrigin-RevId: 514577387
