+++ tensorflow/compiler/xla/hlo/utils/hlo_live_range_test.cc
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow/compiler/xla/service/hlo_live_range.h"
+#include "tensorflow/compiler/xla/hlo/utils/hlo_live_range.h"
 
 #include <memory>
 #include <string>
